# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mang Herbal e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  name: string;
  phone: string;
  address: string;
  /** @nullable */
  note?: string | null;
}